gfx940: Fix the scratch offset calculation
When rocgdb reads from private memory, this warning is reported:
"flat_scratch may be corrupted, private memory access is disabled"

The issue is that the scratch offset calculated by the dbgapi using the
queue's tmpring_size register and the wave's compute_relaunch register
does not match the content of the architected flat_scratch register.

In multi-xcc gfx940, each XCC has its own backing memory for scratch,
so the queue's scratch backing memory which is allocated for all XCCs
should be evenly divided between XCCs.

+++ src/os_driver.h
@@ -75,7 +75,7 @@ struct os_agent_info_t
   /* Number of shader engines.  */
   size_t shader_engine_count{ 0 };
   /* Number of Compute Accelerator Chips.  */
-  size_t xcc_count{ 0 };
+  uint32_t xcc_count{ 0 };
   /* PCI vendor id.  */
   uint32_t vendor_id{ 0 };
   /* PCI device id.  */
